+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG for husky
 
+* dev-master
+    * ENHANCEMENT #605 Added loader to indicate loading suggestions
+
 * 0.18.5 (2016-02-05)
     * HOTFIX      #496 Fixed globalization of datetime value which has a time component in it
     * HOTFIX      #599 Fixed rendering preselected